+++ src/pages/sf-map.tsx
@@ -1,64 +1,19 @@
 import { useState, useMemo } from "react";
-import { Box, Text, Link as ThemeLink } from "theme-ui";
-import Link from "next/link";
+import { Box } from "theme-ui";
 import Head from "next/head";
 import { GetServerSideProps } from "next";
-import Map, {
-  Marker,
-  GeolocateControl,
-  NavigationControl,
-  useMap,
-} from "react-map-gl";
-import { blue, gray } from "@radix-ui/colors";
-import "mapbox-gl/dist/mapbox-gl.css";
+import Map, { GeolocateControl, NavigationControl } from "react-map-gl";
 import { MapboxEvent } from "mapbox-gl";
 
-type PlaceType =
-  | "activity"
-  | "bakery"
-  | "bar"
-  | "cafe"
-  | "dessert"
-  | "other"
-  | "restaurant";
-
-type Record = {
-  id: string;
-  createdAt: string;
-  fields: {
-    name: string;
-    starred: boolean;
-    type: PlaceType;
-    url: string;
-    lat: number;
-    lon: number;
-    price?: string;
-  };
-};
+import "mapbox-gl/dist/mapbox-gl.css";
+import { Record } from "types/sfMap";
+import { CustomMarker, Heading } from "views/sfMap";
+import { mapInitialViewState } from "views/sfMap/constants";
 
 const MAPBOX_TOKEN = process.env.NEXT_PUBLIC_MAPBOX_TOKEN;
 const AIRTABLE_SF_BASE = process.env.AIRTABLE_SF_BASE;
 const AIRTABLE_API_URL = `https://api.airtable.com/v0/${AIRTABLE_SF_BASE}/places?view=app`;
 const AIRTABLE_API_KEY = process.env.AIRTABLE_API_KEY;
-
-// Center view around San Francisco
-const mapInitialViewState = {
-  longitude: -122.45,
-  latitude: 37.778,
-  zoom: 13,
-};
-
-const pinMap = {
-  restaurant: "🍽️",
-  cafe: "☕",
-  dessert: "🍦",
-  bar: "🍸",
-  activity: "📍",
-  bakery: "🥖",
-  other: "📍",
-};
-
-const shadow = "0 1px 3px 0 rgb(0 0 0 / 0.1), 0 1px 2px -1px rgb(0 0 0 / 0.1)";
 
 async function listPlaces(offset?: string) {
   const res = await fetch(
@@ -72,169 +27,6 @@ async function listPlaces(offset?: string) {
   const { records, offset: nextOffset } = await res.json();
   return { records, nextOffset };
 }
-
-const Pin = ({
-  type,
-  isSelected,
-}: {
-  type: PlaceType;
-  isSelected: boolean;
-}) => {
-  const emoji = pinMap[type] || pinMap["other"];
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        justifyContent: "center",
-        alignItems: "center",
-        cursor: "pointer",
-        fontSize: "12px",
-        borderRadius: "100%",
-        background: "white",
-        padding: "2px",
-        width: "24px",
-        height: "24px",
-        boxShadow: shadow,
-        border: isSelected ? `2px solid ${blue.blue7}` : "none",
-      }}
-    >
-      {emoji}
-    </Box>
-  );
-};
-
-const CustomMarker = ({
-  record,
-  onSelect,
-  isSelected,
-}: {
-  record: Record;
-  onSelect: (record: Record) => void;
-  isSelected: boolean;
-}) => {
-  const mapRef = useMap();
-  const handleClick = (mapboxEvent: MapboxEvent<MouseEvent>) => {
-    // If we let the click event propagate to the map, it will immediately close the popup.
-    mapboxEvent.originalEvent.stopPropagation();
-    onSelect(record);
-    if (mapRef?.current) {
-      mapRef.current.flyTo({
-        center: [record.fields.lon, record.fields.lat],
-        duration: 1000,
-        essential: true,
-      });
-    }
-  };
-
-  return (
-    <Marker
-      key={`marker-${record.id}`}
-      longitude={record.fields.lon}
-      latitude={record.fields.lat}
-      anchor="bottom"
-      onClick={handleClick}
-      style={{
-        zIndex: isSelected ? 1 : "auto",
-      }}
-    >
-      <Pin type={record.fields.type} isSelected={isSelected} />
-    </Marker>
-  );
-};
-
-const Popup = ({ selectedPlace }: { selectedPlace: Record }) => (
-  <Box
-    sx={{
-      width: "320px",
-      bg: "white",
-      p: 3,
-      borderRadius: "16px",
-      boxShadow: shadow,
-      pointerEvents: "auto",
-    }}
-  >
-    <Text
-      as="p"
-      sx={{
-        color: gray.gray12,
-        fontFamily: "heading",
-        lineHeight: "heading",
-        fontWeight: "heading",
-        fontSize: [2, 3],
-        letterSpacing: [0, "-0.03em"],
-      }}
-    >
-      {selectedPlace.fields.name}
-    </Text>
-    <Text
-      as="p"
-      sx={{
-        color: gray.gray11,
-        fontFamily: "body",
-        lineHeight: "body",
-        fontWeight: "body",
-        fontSize: 2,
-      }}
-    >
-      {selectedPlace.fields.type}
-    </Text>
-  </Box>
-);
-
-const MapHeading = ({ selectedPlace }: { selectedPlace: Record | null }) => (
-  <Box
-    sx={{
-      position: "absolute",
-      top: 0,
-      zIndex: 9,
-      padding: 4,
-      pointerEvents: "none",
-    }}
-  >
-    <Text
-      as="h1"
-      sx={{
-        display: "block",
-        color: gray.gray12,
-        fontFamily: "heading",
-        lineHeight: "heading",
-        fontWeight: "heading",
-        fontSize: [3, 4],
-        letterSpacing: [0, "-0.03em"],
-        mb: 1,
-        pointerEvents: "auto",
-      }}
-    >
-      San Francisco Food and Fun
-    </Text>
-    <Text
-      as="p"
-      sx={{
-        color: gray.gray11,
-        fontFamily: "body",
-        lineHeight: "heading",
-        fontWeight: 500,
-        fontSize: [2, 3],
-        letterSpacing: [0, "-0.03em"],
-        mb: 4,
-        pointerEvents: "auto",
-      }}
-    >
-      made by{" "}
-      <Link href="/" passHref>
-        <ThemeLink
-          sx={{
-            color: gray.gray11,
-          }}
-        >
-          Andrew
-        </ThemeLink>
-      </Link>
-    </Text>
-    {selectedPlace && <Popup selectedPlace={selectedPlace} />}
-  </Box>
-);
 
 export const getServerSideProps: GetServerSideProps = async (_context) => {
   let locations: Record[] = [];
@@ -292,7 +84,7 @@ const SfMapPage = ({ data }: { data: Record[] }) => {
           display: "relative",
         }}
       >
-        <MapHeading selectedPlace={selectedPlace} />
+        <Heading selectedPlace={selectedPlace} />
         <Map
           initialViewState={mapInitialViewState}
           style={{ width: "100%", height: "100%" }}
