+++ src/views/sfMap/constants.ts
@@ -1,0 +1,9 @@
+// Center view around San Francisco
+export const mapInitialViewState = {
+  longitude: -122.45,
+  latitude: 37.778,
+  zoom: 13,
+};
+
+export const shadow =
+  "0 1px 3px 0 rgb(0 0 0 / 0.1), 0 1px 2px -1px rgb(0 0 0 / 0.1)";
